+++ corvid/ecs/ecs_meta.h
@@ -17,29 +17,20 @@
 #pragma once
 
 #include <cstddef>
-#include <optional>
-#include <tuple>
 #include <type_traits>
+
+#include "../meta/traits.h"
 
 namespace corvid { inline namespace ecs { inline namespace ecs_metas {
 
 // Compile-time helpers shared by the scene aggregators.
 //
-// Provides tuple-membership and indexing metafunctions, the
-// selector-to-storage resolution (`storage_index_for_v`) used by
+// Provides the selector-to-storage resolution (`storage_index_for_v`) used by
 // `archetype_scene` and `component_scene`, and the `get_or_default` migration
-// helper.
+// helper. The generic tuple metafunctions these build on live in
+// "meta/traits.h".
 
 #pragma region Tuple membership
-
-// True if `T` appears at least once in `Tuple`.
-template<typename T, typename Tuple>
-struct tuple_contains;
-template<typename T, typename... Ts>
-struct tuple_contains<T, std::tuple<Ts...>>
-    : std::bool_constant<(std::is_same_v<T, Ts> || ...)> {};
-template<typename T, typename Tuple>
-inline constexpr bool tuple_contains_v = tuple_contains<T, Tuple>::value;
 
 // True if `Storage::tuple_t` contains every type in `Cs...`.
 //
@@ -161,67 +152,6 @@ template<typename C, typename SrcTuple>
   else
     return C{};
 }
-
-#pragma endregion
-#pragma region Tuple operations
-
-// Helper: append T to Tuple only if T is not already present.
-template<typename T, typename Tuple>
-struct tuple_append_unique;
-template<typename T, typename... Ts>
-struct tuple_append_unique<T, std::tuple<Ts...>> {
-  using type = std::conditional_t<(std::is_same_v<T, Ts> || ...),
-      std::tuple<Ts...>, std::tuple<Ts..., T>>;
-};
-
-// Recursively accumulate types from SrcTuples into AccTuple, skipping
-// duplicates. Handles both type-list and empty-tuple cases.
-template<typename AccTuple, typename... SrcTuples>
-struct tuple_union_impl {
-  using type = AccTuple;
-};
-template<typename AccTuple, typename Head, typename... Tail, typename... Rest>
-struct tuple_union_impl<AccTuple, std::tuple<Head, Tail...>, Rest...> {
-  using type =
-      tuple_union_impl<typename tuple_append_unique<Head, AccTuple>::type,
-          std::tuple<Tail...>, Rest...>::type;
-};
-template<typename AccTuple, typename... Rest>
-struct tuple_union_impl<AccTuple, std::tuple<>, Rest...> {
-  using type = tuple_union_impl<AccTuple, Rest...>::type;
-};
-
-// Deduplicated union of component types across all `Tuples`. For example,
-// `tuple_union_t<tuple<A,B,C>, tuple<A,D,E>>` yields `tuple<A,B,C,D,E>`.
-template<typename... Tuples>
-using tuple_union_t = tuple_union_impl<std::tuple<>, Tuples...>::type;
-
-// 0-based index of T in Tuple. Fails to compile if T is not present.
-template<typename T, typename Tuple, size_t I = 0UZ>
-struct tuple_index_impl;
-template<typename T, size_t I>
-struct tuple_index_impl<T, std::tuple<>, I> {
-  static_assert(false, "type not found in tuple");
-};
-template<typename T, typename Head, typename... Tail, size_t I>
-struct tuple_index_impl<T, std::tuple<Head, Tail...>, I>
-    : std::conditional_t<std::is_same_v<T, Head>,
-          std::integral_constant<size_t, I>,
-          tuple_index_impl<T, std::tuple<Tail...>, I + 1>> {};
-
-template<typename T, typename Tuple>
-inline constexpr size_t tuple_index_v = tuple_index_impl<T, Tuple>::value;
-
-// Transform `std::tuple<C0, C1, ...>` to
-// `std::tuple<std::optional<C0>, std::optional<C1>, ...>`.
-template<typename Tuple>
-struct wrap_optionals;
-template<typename... Cs>
-struct wrap_optionals<std::tuple<Cs...>> {
-  using type = std::tuple<std::optional<Cs>...>;
-};
-template<typename Tuple>
-using wrap_optionals_t = wrap_optionals<Tuple>::type;
 
 #pragma endregion
 }}} // namespace corvid::ecs::ecs_metas
